FIXUP: BACKPORT: FROMLIST: kvm/x86: virtual suspend time injection: Implement host side
Create fixup.

+++ virt/kvm/kvm_main.c
@@ -929,7 +929,7 @@ static int kvm_init_mmu_notifier(struct kvm *kvm)
 static int kvm_suspend_notifier(struct kvm *kvm)
 {
 	struct kvm_vcpu *vcpu;
-	int i;
+	unsigned long i;
 
 	if (!virt_suspend_time_enabled(kvm))
 		return NOTIFY_DONE;
@@ -945,7 +945,7 @@ static int kvm_suspend_notifier(struct kvm *kvm)
 static int kvm_resume_notifier(struct kvm *kvm)
 {
 	struct kvm_vcpu *vcpu;
-	int i;
+	unsigned long i;
 
 	if (!virt_suspend_time_enabled(kvm))
 		return NOTIFY_DONE;
